requirements.txt
Django==3.0.7
djangorestframework==3.10
djangorestframework-simplejwt==4.1.2
djoser==2.0.3
django-cors-headers==3.4.0
pillow==7.2.0
